✨ Feat: 상품 도메인 CRUD
- 주석 제거 및 코드 수정.

+++ src/main/kotlin/com/challengeteamkotlin/campdaddy/domain/repository/product/dto/FindBySearchableAndReservationFilterDto.kt
@@ -2,7 +2,7 @@ package com.challengeteamkotlin.campdaddy.domain.repository.product.dto
 
 data class FindBySearchableAndReservationFilterDto(
         val productId:Long,
-        val pricePerDay: Int,
+        val pricePerDay: Long,
         val title:String,
         val content:String,
         val image: String?
